Fix crash when cooking groom data on platform not-supporting strands geometry.
This CL bypass the serialization of strands bulk data when a platform does not support strands geometry, instead of writing dummy data. The reasons behind this is that bulk data are serialized in two step: 1) the initial/regular serialization happens. At that time, the pointer on the bulk data is queued 2) then later the bulk data pointer is dequeued, and the bulk data are serialized. Because dummy resources where stack allocated, their memory was invalid by the time the second serialization step arises.

This CL fixes the issue by not serializating the bulkdata of these dummy resources.

#rb paul.chipchase
#jira none

#ROBOMERGE-SOURCE: CL 16466402 in //UE5/Private-Frosty/... via CL 16466408 via CL 16466411
#ROBOMERGE-BOT: STARSHIP (Release-Engine-Test -> Release-Engine-Staging) (v820-16465646)

[CL 16466414 by charles derousiers in ue5-release-engine-staging branch]