Horde: Fix situations where batches are unnecessarily restarted due to being treated as complete.
Previous method to check whether batch can be modified just checked to see if all steps were complete. This is incorrect, and assumes the batch cannot be updated while the agent is requesting the next batch to execute. Now has an explicit state for checking if the batch is shutting down.

#ROBOMERGE-SOURCE: CL 16506952 in //UE5/Main/... via CL 16507033
#ROBOMERGE-BOT: STARSHIP (Release-Engine-Test -> Release-Engine-Staging) (v826-16501804)

[CL 16507065 by ben marsh in ue5-release-engine-staging branch]