XR: Disable ScreenSpaceDiffuseIndirect rendering for VR
There are some other checkpoints for ShouldRenderScreenSpaceDiffuseIndirect, but they don't trigger ensures. It would be nice to move the multi-view check to ShouldRenderScreenSpaceDiffuseIndirect, but this is probably safer for 4.27.
Also, nobody had tried SSGI with VR until this??

Preflight was clear.

RESOLUTION: Kept target source for UE5 since the core logic is different, and will need time to reconsider. No need to block Robomerge for now.

#rb steve.smith
#jira UE-116781
#lockdown Nick.Whiting

#ROBOMERGE-OWNER: robert.srinivasiah
#ROBOMERGE-AUTHOR: robert.srinivasiah
#ROBOMERGE-SOURCE: CL 16562458 in //UE4/Release-4.27/... via CL 16562465
#ROBOMERGE-BOT: STARSHIP (Main -> Release-Engine-Staging) (v828-16531559)
#ROBOMERGE-CONFLICT from-shelf

[CL 16562542 by robert srinivasiah in ue5-release-engine-staging branch]