Improved and added optional limits around private members to Blueprint thread safety checks
Currently it is not 'safe' for external code/scripts to be able to 'poke' data that is ostensibly thread-safe. Restricting to only private data is one possible option to solve this.
Gated behind a config flag for now as most content does not obey these rules at the moment.

This change also covers a case with 'valid' UObject use on worker threads where those UObjects are returned from thread-safe functions (which can only be native given the BP-level checks). It is assumed that the native thread-safe fucntion returning a UObject 'knows' that the UObject should be safe to access (e.g. some sub-object like a linked anim instance).

#rb Phillip.Kavan

#ROBOMERGE-SOURCE: CL 16569224 in //UE5/Main/... via CL 16569238
#ROBOMERGE-BOT: STARSHIP (Release-Engine-Test -> Release-Engine-Staging) (v828-16531559)

[CL 16569244 by thomas sarkanen in ue5-release-engine-staging branch]