Added support for custom FText::AsDateTime formatting patterns
Added an overload of FText::AsDateTime that will take a custom formatting pattern using strftime-like syntax, rather than only allowing the use of a few hard-coded styles. Internally this pattern is converted into the format used by ICU, which then attempts to find the best matching formatter (so you may not get a perfect match to your requested pattern in all locales).

As FDateTime.ToString() used a non-standard format syntax (which is now documented in the header), there is now a new FDateTime.ToFormattedString() function that uses strftime-like syntax. This avoids breaking backwards compatibility, and is used in builds where ICU is disabled.

Various tests have been added to verify that the format patterns work, and that FText stringification and parsing of the custom variant works as expected.

#jira UE-79857
#preflight 60ef26fe19a0060001555df0
#rb Brooke.Hubert, Austin.Ellis


#ROBOMERGE-SOURCE: CL 16863590 via CL 16863622 via CL 16863626 via CL 16863630 via CL 16863639 via CL 16863646
#ROBOMERGE-BOT: (v836-16769935)

[CL 16863652 by jamie dale in ue5-release-engine-staging branch]