Fix crash when a second light with light shaft bloom enabled is added to the scene.
The bloom history targets were queued for RDG extraction into memory owned by the LightShaftBloomHistoryRTs map. However, this map can be resized while the graph is being built (when a new light is encountered), so the pointers passed to QueueTextureExtraction can become invalid. The solution is to store pointers to FTemporalAAHistory objects in that map, instead of inlining the objects.

#rnx
#rb Will.Damon


#ROBOMERGE-SOURCE: CL 16908536 via CL 16908540 via CL 16908544 via CL 16908554 via CL 16908563
#ROBOMERGE-BOT: (v836-16769935)

[CL 16908564 by mihnea balta in ue5-release-engine-staging branch]