Sequencer: Redesigned the PlayTo API to allow users to specify if the provided time is inclusive or exclusive. This makes the API much more natural to use as you can specify it play to 45 (exclusive) then later jump to 45 and play from there without overlapping evaluations.
Additionally, this fixes some issues with how PlayTo handles the last frame. Previously it would issue a 'Jump' command which skips calling events on the frame being jumped to. But the Jump state also skips sweeping over time ranges. This means that event keys on the last frame of a section (or more depending on framerate) could get skipped, ie:

Playhead is at frame 43. Event key on 44. PlayTo(45). If the delta time was large enough to go from 43 => 45.5 the system would historically issue a Jump command to jump to 45, skipping the event key on 44. It now issues a Play command to cover the gap (which will evaluate the event keys).

The reason a different params struct was added to PlayTo (instead of extending the existing one) is that inclusive/exclusive doesn't make sense for Jump/Seek (which are directionless), and is a params struct so that if future parameters are needed it is easier to add one without breaking blueprints.

#jira UE-119063
#rb Max.Chen
#preflight 610ac2ea5738260001b08767



#ROBOMERGE-SOURCE: CL 17055917 via CL 17056201 via CL 17059271 via CL 17059398
#ROBOMERGE-BOT: (v850-17047176)

[CL 17062802 by matt hoffman in ue5-release-engine-staging branch]