Manual merge of CL 17057436 with some changes to respect UE5 unreal naming changes
Adding ConfigRules option to Enable/Disable/DisableAndStop for Android Background Download. If not set in config rules will treat it as enabled by default.
Added native callbacks for Start/Stop worker in AndroidBackgroundServices and AndroidBackgroundHTTP
Added SetWorkResultOnWorker function to FUEWorkManagerNativeWrapper to call the SetWorkResult java functions on the supplied jobject to simplify this common task when using AndroidBackgroundServices.
Updated UEDownloadWorker calls to always broadcast and then have AndroidPlatformBackgroundHTTPManager decide if it wants to utilize those callbacks. This is a cleaner more consistent approach.
#rb none
#tests Android Development Client


#ROBOMERGE-OWNER: thomas.ross
#ROBOMERGE-AUTHOR: thomas.ross
#ROBOMERGE-SOURCE: CL 17057436 via CL 17058823 via CL 17063029 via CL 17068680 via CL 17068976
#ROBOMERGE-BOT: (v853-17066230)

[CL 17069207 by thomas ross in ue5-release-engine-staging branch]