Fix an issue where characters bumping into physics objects while airborne would stop instead of matching the velocity of the thing they hit.
New CVar p.UseTargetVelocityOnImpact (now enabled by default) changes how characters adjust their velocity when falling and impacting a physics object. Normally, when impacting something during a fall, the character's velocity is recalculated based on the distance they actually moved in the given time. But because physics objects move out of sync with characters, characters act as if they are impacting a stationary object. This CL adds a special case for when impacting physics objects, which calculates the character's new velocity based on the supposed velocity of the contact point on the physics object.

[REVIEW] max.whitehead
[FYI] bill.colby
#tests PIE, two players with a big rig. Confirmed that the new CVar fixes the problem in the jira ticket when enabled.


#ROBOMERGE-SOURCE: CL 17359016 via CL 17359301 via CL 17359329 via CL 17359391 via CL 17359420
#ROBOMERGE-BOT: (v865-17346139)

[CL 17359428 by julian smart in ue5-release-engine-staging branch]