Giving client more authority over jump / landing position to address the sudden correction that happens when the client mounts / dismounts a fast moving base (such as a vehicle or platform).
When a pawn is on a fast moving base, its local-space position may be well synchronised between client and server while its world-space position has a large degree of error. We already favour the local space position, which works great until the client jumps off the fast moving base. When a pawn jumps, their base position is converted to a world position, and suddenly there is a big difference between the client and server pawn location. This normally triggers a network correction for that pawn, suddenly teleporting the client to a new position:
- This is jarring;
- Jumping up and down on a moving base consistenly causes two corrections (take off and landing);

However, if we defer the correction until the client has landed, their jump is no longer jarring, and jumping up and down on a moving base causes no corrections. BUT, landing on a different base moving at a different velocity (or not at all) will trigger a correction that's jarring for the client, who may have been quite intentional about where they thought they were landing. To combat that issue, we can allow the client a little bit of authority over landing position -- when landing, before the server sends a correction to the client, the server first updates its own pawn representation's location to better match the client location (up to a threshold). To avoid trusting the client too much, the logic and conditions to enable this are all handled by the server.

So here are our new CVars:
- p.ClientAuthorityThresholdOnBaseChange (recommend 150cm) - How far the server is willing to move towards the client position upon landing. Trusting the client makes it easier for them to land on differently moving objects, with the trade-off being that observers may see a correction on landing.
- p.MaxFallingCorrectionLeash (recommend 150cm) - When a pawn jumps off a moving base, we remember how far apart the client and server representations are and don't allow that error to increase. While airborne, when this limit is exceeded, the client will be moved just enough to be within that limit rather than all the way to the server position. This limit only applies while the pawn is airborne. Regular corrections apply again when the pawn is on the ground (after the above-mentioned correction is applied upon landing). This CVar "MaxFallingCorrectionLeash" is the maximum allowable error for airborne pawns. So with the default value, even if the client and server are 200cm apart when they first leave their moving base, the client will be immediately corrected to within 150cm of the server location.
- p.MaxFallingCorrectionLeashBuffer (default 10cm) - When correcting airborne pawns with the leash value described above, pull the client in even further by this amount to avoid having to make slight corrections over and over when the client is just on the edge of the correction threshold. For example, with the default values, if an airborne client is greater than 150cm from the server representation, it'll be pulled to (150 - 10 =) 140cm of the server representation.

Setting any of these CVars to 0 disables that part of this change. Setting them all to 0 will make it as if this change was never submitted. Spooky!

[REVIEW] [at]max.whitehead [at]bill.colby
#tests PIE, three players (KBM and controller) with sport, truck, bigrig, lots of jumping up and down, running around while the other drives. Screen capture to look back and see how it looks from both the player on top of the vehicle and the player driving.


#ROBOMERGE-OWNER: julian.smart
#ROBOMERGE-AUTHOR: julian.smart
#ROBOMERGE-SOURCE: CL 17378929 via CL 17379125 via CL 17379156 via CL 17380845 via CL 17380846
#ROBOMERGE-BOT: (v865-17346139)

[CL 17380847 by julian smart in ue5-release-engine-staging branch]