Bug fix for startup OOM in trace logging due to infinite allocation loop.
The padding calculation in the function that allocates buffers for trace logs (Writer_CreateSharedBuffer) was off by 4 bytes.  The function is called by a lock free allocator which loops until an allocation succeds that's large enough to fit the requested size.  Because of the calculation error, allocation sizes between 997 and 1000 bytes would result in a block being allocated that was too small, so it would never succeed, and retry the allocation an infinite number of types.

Other sizes would work fine, because there is padding to a multiple of the block size.  I was personally seeing the error specifically because my command line (part of the information written to the log) when running a package build was a precise length that resulted in a size of 1000 bytes when added to a bunch of other strings written to the log.  Besides fixing the bug, I added some comments explaining what the lock free allocator is doing.

#rb Martin.Ridgers Brandon.Dawson
#rnx
[FYI] Jason.Hoerner

#ushell-cherrypick of 17303709 by Jason.Hoerner
#preflight 612f364775bca20001f22c05



#ROBOMERGE-SOURCE: CL 17384407 via CL 17384417 via CL 17384420 via CL 17384421
#ROBOMERGE-BOT: (v865-17346139)

[CL 17384424 by martin ridgers in ue5-release-engine-staging branch]