Alembic: improve build/install/deploy scripts on Windows
In earlier versions, the HDF5 and Alembic build scripts took care of the build
and install steps, but that left the build products in a state where they still
needed to be rearranged manually to remove unnecessary files and organized
into the directory structure that the Alembic.Build.cs expected.

With these changes, HDF5VS.bat and AlembicVS.bat now take care of that
final deployment step as well. They will also handle removing the existing
deployment beforehand. The build, installation, and deployment of each of
HDF5 and Alembic are now fully encapsulated in their respective scripts, and
they invoke the compiler through CMake rather than by directly calling MSBuild.
The "top-level" abc.bat script now simply runs HDF5VS.bat followed by AlembicVS.bat.
If run independently, AlembicVS.bat will verify that it is only run after a successful
run of HDF5VS.bat, since the Alembic build depends on HDF5 being available in
the typical installation directory structure where the "include" and "lib" directories
are siblings.

Finally, the files generated by the build and install steps are created in the
"Intermediate" directory of the module to make clear that they are fully disposable
and safe to remove.

Going through this process helped detect that the main hdf5.h header file was not
being deployed from the installation, so although it is not currently used, it is added
here for completeness.

#rb anousack.kitisa

#ROBOMERGE-AUTHOR: matt.johnson
#ROBOMERGE-SOURCE: CL 17463680 in //UE5/Main/... via CL 17463766
#ROBOMERGE-BOT: STARSHIP (Release-Engine-Test -> Release-Engine-Staging) (v870-17433530)

[CL 17463822 by matt johnson in ue5-release-engine-staging branch]