#jira UE-121870 Datasmith Tessellation does not work as operator in the visual dataprep
FCoreTechInterfaceImpl::RepairInternal create a filter to process only bodies that need to be repaired.
In this jira case, the body has not problem. So the part is not in the filter set.
All KernelIO function use the filter. CT_KERNEL_IO::SaveFile too. So the body is not saved.
At the end of the process, filter is deleted. That fix the jira

#rb jeanluc.corenthin, Emil.Kirichev
#lockdown Mitchell.Wilson

#ROBOMERGE-AUTHOR: david.lesage
#ROBOMERGE-SOURCE: CL 17475724 in //UE4/Release-4.27/... via CL 17475767 via CL 17475770
#ROBOMERGE-BOT: STARSHIP (Main -> Release-Engine-Staging) (v870-17433530)

[CL 17475791 by david lesage in ue5-release-engine-staging branch]