Added an interface class to Gauntlet to represent TestEvents (ITestEvent) that occur during testing. Events are intended to replace the use of strings for GetErrors/GetWarnings in tests nodes to generalize the approach of UE.Automation where events are structured objects. Once done UE.Automation will be closer to UnrealTestNode and much of the custom parsing/logging code can be removed.
The immediate benefit is that Unreal tests can  specify a set of log categories where warnings/errors should be treated as warnings/errors in the test. This removes the need for many tests to override reporting functions and use custom log parsing to report that, for example, a ClientBot reported error "Foo".

This is super valuable for games where "normal behavior" can result in warnings or errors in certain log channels. Tests can specify a set of categories that should be error free during normal operation and these will be surfaced automagically with no need for custom parsing or log reading.


public override FortTestConfig GetConfiguration()
{
     FortTestConfig Config = base.GetConfiguration();

    // Mark warnings/errors in these categories as being akin to warning/error test events.
    // E.g. a LogError in FortClientBot will be reported as a test error
    Config.LogCategoriesForEvents.AddRange(new string[]  {
        "Init",
        "GameFeatures",
        "FortClientBot"
    });
<snip>


#rb swarm
#preflight https://horde.devtools.epicgames.com/log/613be6af85759000015dd12b
[FYI] jerome.delattre


#ROBOMERGE-OWNER: andrew.grant
#ROBOMERGE-AUTHOR: andrew.grant
#ROBOMERGE-SOURCE: CL 17344222 via CL 17344855 via CL 17484529 via CL 17484545
#ROBOMERGE-BOT: (v870-17433530)

[CL 17484546 by andrew grant in ue5-release-engine-staging branch]