Fix for some FN tests failing with UE5
There's a UE5 change that results in ScanPathsSynchronous accessing an invalid array index if empty paths are passed in. This could be considered invalid, but delegates are (I think) fired by the asset registry so just early-out'ing with no action might be problematic. As a result I modified the code that logs the paths to handle this.

On the other hand I don't see any need for the Fortnite code to perform this scan if a plugin has no mount points when it's unloaded so I've also modified that code to skip it.

[at]Matt.Peters [at]Dave.Belanger
#preflight https://horde.devtools.epicgames.com/job/613d4fd472f6640001bbfb28?label=18


#ROBOMERGE-AUTHOR: andrew.grant
#ROBOMERGE-SOURCE: CL 17487147 via CL 17487149
#ROBOMERGE-BOT: (v870-17433530)

[CL 17487151 by andrew grant in ue5-release-engine-staging branch]