Undo changes to transition vertex buffers each time they are bound (CL 17489558).
Unfortunately the state cache can contain dangling pointers to buffers which have been freed or renamed mid-frame, because nothing clears slots in the vertex buffer cache until the end of the frame. This was "fine" before, since we only touch the views in SetVertexBuffers, and as long as the shader doesn't use the bogus slots, nothing bad happens. However, the change made that function touch CurrentVertexBufferResources, which can contain pointers to freed memory.

This brings back the problem that the code can miss transitions. We'll have to revisit this and properly clear the cache when resources are deleted or renamed.

#rb trivial
#rnx
#jira UE-126601

#ROBOMERGE-AUTHOR: mihnea.balta
#ROBOMERGE-SOURCE: CL 17519494 in //UE5/Main/... via CL 17519507
#ROBOMERGE-BOT: STARSHIP (Release-Engine-Test -> Release-Engine-Staging) (v870-17433530)

[CL 17519517 by mihnea balta in ue5-release-engine-staging branch]