Revisited the fix made to prevent packaged blueprint games by a vanilla Engine downlowded from the store to report analytic and bug report.
The fix made at CL 15279807 prevented internal products to send unattended bug reports. A workaround was implemented at CL 17281496 and 17281531, but that implies that all internal products needed to add a new define, which is not  intuitive and hard to discover. It was better to revert the change and reintroduce the original bug because this was less of a problem. At the moment, there is no way to distinguish from a blueprint only game packaged from a vanilla engine versus a internal product without introducing a new global define.

#jira UE-83082 - Ensure events are firing in packaged games packaged with default settings and Crash Report Client included

#rb none


#ROBOMERGE-AUTHOR: patrick.laflamme
#ROBOMERGE-SOURCE: CL 17543369 via CL 17543421 via CL 17543422
#ROBOMERGE-BOT: (v870-17433530)

[CL 17543425 by patrick laflamme in ue5-release-engine-staging branch]