Fix crash when a second light with light shaft bloom enabled is added to the scene.
The bloom history targets were queued for RDG extraction into memory owned by the LightShaftBloomHistoryRTs map. However, this map can be resized while the graph is being built (when a new light is encountered), so the pointers passed to QueueTextureExtraction can become invalid. The solution is to store pointers to FTemporalAAHistory objects in that map, instead of inlining the objects.

#jira FORT-385825, UE-115699
#rnx
#rb Will.Damon
#lockdown Michal.Valient

#ushell-cherrypick of 16908536 by mihnea.balta

#ROBOMERGE-AUTHOR: mihnea.balta
#ROBOMERGE-SOURCE: CL 17606212 in //UE4/Release-4.27/... via CL 17606237 via CL 17606246 via CL 17614939
#ROBOMERGE-BOT: STARSHIP (Release-5.0 -> Release-Engine-Staging) (v871-17566257)

[CL 17615068 by mihnea balta in ue5-release-engine-staging branch]