Better movement when override root motion moves a character off a moving base.
Adds a new property to CharacterMovementComponent: FormerBaseVelocityDecayHalfLife.

When moving off from on top of a moving base, that base's velocity is imparted to the pawn to respect inertia. But if the pawn has a root motion source set to "override", that inertia is gone instantly. FormerBaseVelocityDecayHalfLife controls the rate at which that inertia disappears. Its default value of 0 behaves the same as before this change -- inertia is gone instantly. A value of 1 means it'll take 1 second to lose half the pawn's inertia. 2 will take 2 seconds to lose half its inertia, and so on. To maintain inertia with no decay for as long as the pawn is falling, set FormerBaseVelocityDecayHalfLife to -1.

This new behavior can be disabled by setting the CVar p.AddFormerBaseVelocityToRootMotionOverrideWhenFalling to '0'.

This CL also fixes former base movement not being retained correctly when applying "additive" root motion.

[REVIEW] max.whitehead
[FYI] bill.colby
#tests PIE



[CODEREVIEW] Justin.Hare, Zak.Middleton, Julian.Smart, Max.Whitehead

#ROBOMERGE-OWNER: marc.audy
#ROBOMERGE-AUTHOR: julian.smart
#ROBOMERGE-SOURCE: CL 17603878 via CL 17603885 via CL 17603889 via CL 17603891 via CL 17603895 via CL 17615358
#ROBOMERGE-BOT: STARSHIP (Release-5.0 -> Release-Engine-Staging) (v871-17566257)

[CL 17615399 by marc audy in ue5-release-engine-staging branch]