Sequencer: Custom Curve Colors
Added Custom Color Support to Curve Editor. This is done via an Editor preference(CurveEditorSettings) that stores a color per UClass, property name(intent name of the channel in Curve Editor), and the Color. So would be something like StaticMeshActor, "Translate.X", (1.0f,0.0f,0.f90)

Added to Sections an UObject* GetImplicitObjectOwner(). Usually this is the Possessable/Spawnable of the owning track, but for Control Rig Tracks we use the
specified Control Rig instead.

So we use that implicit owner to get the UClass to associate with the channel.

Also added some Python/BP functions since editor pref's aren't scriptable.

#jira MH-3206
#preflight 61520812d0edce00010149bf

#ROBOMERGE-AUTHOR: mike.zyracki
#ROBOMERGE-SOURCE: CL 17635750 in //UE5/Main/... via CL 17635764 via CL 17635771
#ROBOMERGE-BOT: STARSHIP (Release-Engine-Staging -> Release-5.0) (v871-17566257)

[CL 17635778 by mike zyracki in 5.0 branch]