The DXGI factory pointer stored by the D3D11 RHI has to be the same one that was used to create the adapter.
The code was creating one DXGI factory when it made the device, but then it made a second factory which was stored in FD3D11DynamicRHI and used for subsequent calls. This resulted in a subtle problem with swapchain creation: CreateSwapChain was called on the second factory pointer, but internally it doesn't care and uses the parent factory of the device, i.e. the first factory we were making. When we called MakeWindowAssociation on the same (second) pointer, it didn't work, because it was not the factory which really owned the swapchain. This meant that DXGI was still hooking our window proc and catching alt+enter and tried to handle fullscreen for us, which broke all sorts of things.

#rnx
#lockdown Michal.Valient
#rb Christopher.Waters

#ushell-cherrypick of 17642278 by mihnea.balta

#ROBOMERGE-AUTHOR: mihnea.balta
#ROBOMERGE-SOURCE: CL 17642757 via CL 17642931 via CL 17643005 via CL 17643007
#ROBOMERGE-BOT: STARSHIP (Release-5.0 -> Release-Engine-Staging) (v875-17642767)

[CL 17643021 by mihnea balta in ue5-release-engine-staging branch]