Send buffers' data closer to age than by thread.
Trace's normal-event buffers form a list by thread, and for each thread there is a list of its buffers. Prior to this change, the entire buffer list for a thread was sent out. Now just one buffer is sent, and the threads requeued if there's more buffers to go. This multiplexes the threads more in the stream of traced data. Consecutive events are less likely to be so far apart and tails represent more threads.

#rb jb
#jira ue-127376
#rnx
#preflight 6152dae16ca13700011f8b5e

#ROBOMERGE-AUTHOR: martin.ridgers
#ROBOMERGE-SOURCE: CL 17643566 in //UE5/Main/... via CL 17643569
#ROBOMERGE-BOT: STARSHIP (Release-Engine-Test -> Release-Engine-Staging) (v875-17642767)

[CL 17646431 by martin ridgers in ue5-release-engine-staging branch]