Tasks: removed invalid `verify` as it's triggered in an valid case:
Thread1 can call TaskEvent.Trigger(), which unlocks the task, and before it calls verify(TryExecute()), Thread2 starts TaskEvent retraction (because Thread2 started waiting on its subsequent task), and first it gets execution permission. then it proceeds to complete retraction as the task is unlocked but not executed yet. Now Thread1 tries to get execution permissions and fails, but can proceed as normal. its job is done

#[fyi] devin.doucette
#[robomerge] 5.0

#ushell-cherrypick of 17770492 by Andriy.Tylychko


#ROBOMERGE-AUTHOR: andriy.tylychko
#ROBOMERGE-SOURCE: CL 17782609 via CL 17782611 via CL 17782614 via CL 17782617
#ROBOMERGE-BOT: STARSHIP (Release-5.0 -> Release-Engine-Staging) (v881-17767770)

[CL 17782629 by andriy tylychko in ue5-release-engine-staging branch]