StallDetector: Support for pausing stall timers
### Features

SCOPE_STALL_DETECTOR_PAUSE() will prevent the stall detector thread of final synchronous check from issuing a stall for inclusive time in the scope

### Notes

Previously, StartSeconds only tracked when the StallDetector was constructed, and allowed for stall checks to be immutable relative to StallDetector objects (we didn't need to modify the StallDetector to check the transpired time vs. budget).

This can no longer be true if we want to allow pausing, some writes will be issued during a stall check (and from two threads that could perform a check: the potentially stalling thread w/ the detector, or the StallDetector thread).

So, create a new FStallTimer object to store and protect LastCheckSeconds (which gets set every check) and RemainingSeconds (which keeps a running tally of the allowed time remaining from the original budget). Contention should be low, and we only perform checks every 5ms, so any waiting shouldn't be too costly (fastpath should be in effect almost all the time).

This change also more clearly communicates if we are actually detecting or pausing from the ctor to the dtor of FStallDetector (w/ bStarted) and FStallDetectorPause (w/ bPaused). These bools are better than checks on the timer because it would need to acquire the lock to read that state.

### Testing

Abused FEngineLoop::Tick() with some heinous Sleeps inside nested pause macros inside nested stall detectors
Opened a FMessageDialog in FEngineLoop::Tick() and saw no stalls reported
Lowered the stall timer in FEngineLoop::Tick() and verified it was able to report a stall outside pause macro scope

### Tags

#jira UETOOL-4459
#rb francis.hurteau
#preflight 6169f8e79390dd0001c94f54

#ROBOMERGE-AUTHOR: geoff.evans
#ROBOMERGE-SOURCE: CL 17867679 in //UE5/Release-5.0/...
#ROBOMERGE-BOT: STARSHIP (Release-5.0 -> Release-Engine-Staging) (v883-17842818)

[CL 17867692 by geoff evans in ue5-release-engine-staging branch]