Fix skylight intensity not updating through sequencer
The skylight intensity and color is baked as part of the skydome importance tables for the path tracer, RT skylight and RT reflections. If this change is done without creating a new skylight proxy (for example when keyframed via the sequencer) it is possible the change to go un-noticed. For a similar reason, the path tracer would not restart sample accumulation in this case.

The fix is to just store the color used the last time the importance tables were baked and detect when this has changed. An alternative fix would be to leave the multiplication by the skylight color outside of the importance tables, but it seems better to fold this in as skylight intensity changes are probably rare.

#jira UE-133028
#rb Juan.Canada
#preflight 618326af92492200018b5559

#ROBOMERGE-AUTHOR: chris.kulla
#ROBOMERGE-SOURCE: CL 18044643 in //UE5/Main/... via CL 18044735
#ROBOMERGE-BOT: STARSHIP (Release-Engine-Test -> Release-Engine-Staging) (v885-17909292)
#ROBOMERGE[STARSHIP]: Release-5.0

[CL 18044742 by chris kulla in ue5-release-engine-staging branch]