UE5 + Instanced Stereo Rendering: Skip check for re-used LoadBalancers in ShadowProjection pass
FProjectedShadowInfo::RenderProjection will re-use the ProjectionStencilingPass for the left and right eye. When we go through the right-eye path when ISR is enabled, we re-use the contained FInstanceCullingContext. FInstanceCullingContext::SetupDrawCommands doesn't expect to be re-used, and fails a check where the LoadBalancers are empty.

I can skip the check if ISR is enabled, and the results seem correct.

Future-looking: support ISR natively on shadow projections

#jira UE-123704
#rb ola.olsson jules.blok steve.smith

#ROBOMERGE-AUTHOR: robert.srinivasiah
#ROBOMERGE-SOURCE: CL 18141418 in //UE5/Release-5.0/...
#ROBOMERGE-BOT: STARSHIP (Release-5.0 -> Release-Engine-Staging) (v889-18060218)

[CL 18141440 by robert srinivasiah in ue5-release-engine-staging branch]