UBT: Fix engine failing to build on startup when a project module is missing.
WriteMetadataMode is now run twice - once to update the engine *.modules files and the *.version file, and a second time to update the target-specific *.modules files and receipt. As a result, the engine action does not need to execute unless engine files have changed.

The tool still attempts to preserve the BuildId in the *.version file if the existing files are still compatible (ie. an almost-null build, since the check happens when the action is executed rather than checked for up-to-dateness).

#jira UE-119924

#ROBOMERGE-AUTHOR: ben.marsh
#ROBOMERGE-SOURCE: CL 18238182 in //UE5/Release-5.0/...
#ROBOMERGE-BOT: STARSHIP (Release-5.0 -> Release-Engine-Staging) (v895-18170469)

[CL 18238212 by ben marsh in ue5-release-engine-staging branch]