Transition vertex buffers right before calling IASetVertexBuffers, instead doing it when the high level calls RHISetStreamSource.
If a command list flush happens in the middle of a sequence of draws which use the same vertex buffer, the tracked internal state of the buffer is reset. Since the high level doesn't call RHISetStreamSource again after the flush, no pending barrier was inserted. If a subsequent dispatch needed the buffer in a different state, that transition was erroneously inserted in the patch-up command list, which modified the state too soon (before the draws which still need the buffer in the VB state). Adding the transition in FD3D12DescriptorCache::SetVertexBuffers ensures that the correct state is applied, since this method is called internally before each draw.

Resubmit with fixes for level reload caued by dangling pointers in the state cache: state cache is cleared during calls to RHISubmitCommandsHint when scene or viewport rendering is not enabled (eg level loading)

#jira UE-135116
#rb Mihnea.Balta

#ROBOMERGE-AUTHOR: kenzo.terelst
#ROBOMERGE-SOURCE: CL 18280860 in //UE5/Release-5.0/...
#ROBOMERGE-BOT: STARSHIP (Release-5.0 -> Release-Engine-Staging) (v895-18170469)

[CL 18280866 by kenzo terelst in ue5-release-engine-staging branch]