MovieScene: deprecate UMovieSceneSequenceExtensions::GetTimecodeSource() in favor of GetEarliestTimecodeSource() on UMovieScene and UMovieSceneSequence
UMovieScene is a container that may represent multiple pieces of media as
sections, each of which may have their own distinct timecode source. As a result,
the timecode source associated with the UMovieScene itself doesn't correspond to
any particular piece of media, but rather was just intended as the timecode at which
the movie scene was recorded.

This is the first in a series of changes that aim to remove the TimecodeSource property
of UMovieScene to try to disambiguate it from the similarly named property on
UMovieSceneSections that *does* correspond to the section's media. A subsequent change
will remove the property from UMovieScene, possibly replacing it with one that more
explicitly represents the time at which the movie scene was recorded.

UMovieSceneSequenceExtensions::GetTimecodeSource() previously would return the
TimecodeSource of its UMovieScene. With that property going away, the more interesting
query is likely to look through the media contained by the movie scene and return the
earliest timecode source of any of the scene's sections. A GetEarliestTimecodeSource()
function was added to UMovieScene that does this and the UMovieSceneSequenceExtensions
function was deprecated. A GetEarliestTimecodeSource() function was added to
UMovieSceneSequence as well that simply calls through to its movie scene.

These changes also ensure that the Timecode property of FMovieSceneTimecodeSource
is accessible from Python.

#rb max.chen, jason.walter
#preflight 61b938864331fd305c8e9345

#ROBOMERGE-AUTHOR: matt.johnson
#ROBOMERGE-SOURCE: CL 18464038 in //UE5/Release-5.0/...
#ROBOMERGE-BOT: STARSHIP (Release-5.0 -> Release-Engine-Staging) (v899-18417669)

[CL 18464040 by matt johnson in ue5-release-engine-staging branch]