Takes: set TimecodeSource on UMovieSceneTakeSection when processing a recorded take
When recording a take with timecode in Take Recorder, the UMovieSceneTakeSection
that stores the timecode data is not created until
UTakeRecorderActorSource::ProcessRecordedTimes(), which happens when recording
stops. It therefore isn't available at the start of recording to have its timecode source
set with all of the other movie scene sections for the take. Previously, this was leaving
the TimecodeSource property of the UMovieSceneTakeSection in its default state,
inconsistent with the property values of all other sections for the take.

This adds handling at the end of ProcessRecordedTimes() to ensure that the TimecodeSource
property is set correctly on the take section, either re-using the collected timecode data
from recording, or by adopting the TimecodeSource of another one of the take's sections
if that timecode data is not available for some reason.

#rb max.chen
#preflight 61de3806cd2cced71075fed4

#ROBOMERGE-AUTHOR: matt.johnson
#ROBOMERGE-SOURCE: CL 18586551 in //UE5/Release-5.0/...
#ROBOMERGE-BOT: STARSHIP (Release-5.0 -> Release-Engine-Staging) (v899-18417669)

[CL 18586575 by matt johnson in ue5-release-engine-staging branch]