nDisplay: Fix lightcards not showing when they use a translucent material.
The issue was that the EngineShowFlags passed to the ViewFamily was not yet configured, and instead it gets configured after view family construction. In particular, the separate translucency flag has a side effect during view family construction of setting its internal bAllowTranslucencyAfterDOF flag, which would be missed if SetSeparateTranslucency is called later. In such case, translucency goes into a separate render target and is not captured as expected.

So the fix consists in configuring the EngineShowFlags first, and then passing them to the view family constructor via the FSceneViewFamily::ConstructionValues structure.

#rnx
#jira UE-140072
#rb peter.tarasenko
#preflight 61f2d1a76b5aea38e588d4d8

#ROBOMERGE-AUTHOR: alejandro.arango
#ROBOMERGE-SOURCE: CL 18755139 in //UE5/Release-5.0/...
#ROBOMERGE-BOT: UE5 (Release-5.0 -> Release-Engine-Staging) (v903-18687472)

[CL 18755146 by alejandro arango in ue5-release-engine-staging branch]