Control Rig: change the SpacePicker window to be a popup window so that it automatically closes when user clicks somewhere else.
The old approach causes infinite recusion because RequestDestroyWindow would trigger WindowDeactivate, and the OnDeactivated delegate we use would then invoke closedialog() (namely, RequestDestroyWindow) again. It surprisingly worked on Windows, but did not work on Mac, which makes sense.

Current approach: if a space is picked, closedialog is triggered to close the window proactively. if user clicks outside of the window, normal popup window destruction path is invoked.

#jira UE-137939
#rb trivial
[FYI] sara.schvartzman, helge.mathee, andrew.porter
#preflight skip

#ROBOMERGE-AUTHOR: jack.cai
#ROBOMERGE-SOURCE: CL 18793946 in //UE5/Release-5.0/...
#ROBOMERGE-BOT: UE5 (Release-5.0 -> Release-Engine-Staging) (v908-18788545)

[CL 18794655 by jack cai in ue5-release-engine-staging branch]