Fixed the typed element scripting api issues.
Added the FScriptTypedElementHandle. These handles are disarmed when the element they pointing to is destroyed instead of crashing the engine. They do have a performance overhead so using these should be restricted to exposing stuff to blueprint/python.

Reworked the TypedElementList to be a template so that we have both a FTypedElementList and a FScriptTypedElementList from the same source code.

Changed the api of the interfaces so that can now accept and use the scripted version of the handle and list instead of the native ones.

#jira UE-133667
#rb Brooke.Hubert
#preflight 61f89bfaf657e25a5908db48

#ROBOMERGE-AUTHOR: julien.stjean
#ROBOMERGE-SOURCE: CL 18816318 in //UE5/Release-5.0/...
#ROBOMERGE-BOT: UE5 (Release-5.0 -> Release-Engine-Staging) (v908-18788545)

[CL 18816336 by julien stjean in ue5-release-engine-staging branch]