Duplicating CL#18549727 from //UE5/Release-5.0
#jira UE-138657
CookRequestClusters: suppress the timeout warning on reading vertex data if we're not reading vertexes from the oplog. In the non-oplog case, the vertex data is empty and is just passed through on an async thread rather than waiting for the oplog. It's unknown why that async thread could take up to 10 seconds (hitting a breakpoint maybe), but we don't need a warning about it.
#rb None, Trivial
#rnx
#preflight 61d89964c65e66487b381d9c


#ROBOMERGE-AUTHOR: bob.tellez
#ROBOMERGE-SOURCE: CL 18809938 via CL 18809954 via CL 18809969
#ROBOMERGE-BOT: (v908-18788545)

[CL 18821435 by bob tellez in ue5-release-engine-staging branch]