UE-141320 - Float to double conversion warning in existing blueprint function
Macro assignment nodes don't work like other BP nodes: only their nets refer to meaningful data. As a result, its RegisterNets function needs to insert a custom entry into the implicit cast table.

However, when the compiler calls RegisterImplicitCasts to detect implicit cast points in the graph, it can mistakenly identify that a cast is needed for its value pin. This is incorrect since a cast would only be needed between the *nets* of the value and variable pins. The fix here is to have the node explicitly remove the erroneous entry from the map, which we do in other nodes that have special semantics.

Additionally, changed the math expression handler to use RemoveRegisteredImplicitCast, which is just a simple wrapper around ImplicitCastMap.Remove that returns the number of entries removed. This can be useful for debugging when trying to determine how many entries were removed from the cast table.

#rb dave.jones
#jira UE-141320
#preflight 61fd965ff370b0d3111f7155
#lockdown julien.marchand

#ROBOMERGE-AUTHOR: dave.jones2
#ROBOMERGE-SOURCE: CL 18875868 in //UE5/Release-5.0/...
#ROBOMERGE-BOT: UE5 (Release-5.0 -> Release-Engine-Staging) (v910-18824042)

[CL 18876118 by dave jones2 in ue5-release-engine-staging branch]