Chaos Cloth - updates to PBDCollisionSpringConstraintsBase (cloth self collisions) to improve stability by using simple proximity detection and point-based repulsion.
This required a proximity detection method for cloth. Currently proximities are only detected as point-face. This was done by adding a TAABBTree to the TriangleMesh, and adding a PointProximityQuery. I'm not sure if TriangleMesh is the best place for the BVH to live since the TriangleMesh does not own the point positions, but it seemed like a reasonable place to start. I opted to put it here rather than inside the collision springs class directly because I believe the BVH will be useful for other cloth collision operations in the future (e.g., intersection detection).

Additionally, I removed some data for self collisions in PBDEvolution that is unused.

==CHANGES==
PBDCollisionSpringConstraintsBase.{h,cpp}
This now holds a reference to the TriangleMesh and optional RestPositions. The TriangleMesh is used for its topology and the BVH. The RestPositions are used to exclude collisions between points that are closer than thickness away in the rest state. Normals was removed since it's not longer used.
Init (the method that creates the constraints) now uses the BVH's PointProximityQuery and creates up to 3 constraints per point. This leads to a modest reduction in jitter as connections tend to be more stable compared with the previous more prediction-based detection method.
GetDelta has been updated to be a simple point repulsion rather than the triangle normal-based repulsion from before.  This was the most significant change when it comes to jitter. There was a bug in the previous method that caused too large of deltas in some cases, but even fixing that bug did not resolve the jitter problems.

TriangleMesh.{h,cpp}
Added the BVH here and added methods to update the BVH and do PointProximityQuery

Plane.h
Added new method FindClosestPointAndBaryOnTriangle. This should produce the same closest point as FindClosestPointOnTriangle (up to floating point differences), but it has less duplicate work, and returns the calculated barycentric coordinates. It also handles degenerate triangles like the old code in PBDCollisionSpringConstraintsBase.

PBDCollisionSpringConstraints.h, ChaosClothConstraints.{h,cpp}, ChaosClothingSimulationCloth.cpp
Updated with new constructor for PBDCollisionSpringConstraintsBase
Modified the magic number multiplier in Apply to .5. This worked better with the new constraint delta.

ChaosClothingSimulation.cpp
Updated debug draw for collision springs.

ChaosClothingSimulationCloth.cpp
Updated FLODData to manage the BVH in the triangle mesh, including a "PreTimestep" which rebuilds the BVH.

ChaosClothingSimulationSolver.cpp
Added calling PreTimestep on the cloth.

PBDEvolution.{h,cpp}
Removed unused self collision data.

TriangleCollisionPoint.h
New struct for holding triangle based collision data.

==TESTING==
I tested by enabling self collisions in the example in UE-104875 as well as in the Cloth map in ContentExamples. Jitter is almost completely gone in these examples. Cloth collisions can still occur for fast moving cloth. The topline simulation time for cloth was not significantly different between this and the previous self collision method.

#rb kriss.gossart, benn.gallagher
#jira UE-104875
#preflight 6201688a4f30e5369a480ceb
#lockdown michael.lentine

#ROBOMERGE-AUTHOR: alex.mcadams
#ROBOMERGE-SOURCE: CL 18891743 in //UE5/Release-5.0/...
#ROBOMERGE-BOT: UE5 (Release-5.0 -> Release-Engine-Staging) (v910-18824042)

[CL 18892103 by alex mcadams in ue5-release-engine-staging branch]