Direct Link in Editor - Fixing crash happening when auto-reimport is triggered while we are auto-saving.
The auto-save's "FSlowTask update" triggers a Slate update which can in turn execute pending task on the MainThread. This could cause a DirectLink auto-reimport task to execute while saving, causing a crash. To fix this we now make sure that we are in a valid context before triggering a reimport.

#jira UE-141835
#rb JeanLuc.Corenthin
#preflight 6202f8207e40979d33bcc403
#lockdown JeanMichel.Dignard

#ROBOMERGE-AUTHOR: benoit.deschenes
#ROBOMERGE-SOURCE: CL 18918649 in //UE5/Release-5.0/...
#ROBOMERGE-BOT: UE5 (Release-5.0 -> Release-Engine-Staging) (v916-18915374)

[CL 18926728 by benoit deschenes in ue5-release-engine-staging branch]