Fix visualization LOD Multithreading issue
Since we do not have other ways to create data dependency, we need to create different processors to have independent dependencies (crowd, Traffic, Intersection are ran in parallel). Doing that does not work well with the current implementation where we iterate on all shared fragments of a certain type, because we will iterate on all 3 types of shared fragment in all 3 derived processor class. If the 3 processors are ran in parallel, this is where the problem happens. By specifying a Filter tag in the shared fragment to match the one in the processor so that it only handle the matching tag, it fixes the problem but not it is not an ideal fix.

#jira UE-143435
#rnx
#rb mieszko.zielinski
#preflight 6217d35dbd28238bde8c6000
#lockdown julien.marchand

#ROBOMERGE-AUTHOR: maxime.mercier
#ROBOMERGE-SOURCE: CL 19121884 in //UE5/Release-5.0/...
#ROBOMERGE-BOT: UE5 (Release-5.0 -> Release-Engine-Staging) (v921-19075845)

[CL 19122488 by maxime mercier in ue5-release-engine-staging branch]