Fix right-eye flickering to black when using the SteamVR plugin.
The compositor caches the last result of GetProjectionRaw(), but instead of caching internal values it caches the contents of the output values. By setting the output values to the same memory locations the field-of-view would effectively become zero. Which results in a black screen whenever the compositor uses these field-of-view values to reproject a late frame.

#jira UE-145085
#preflight 623243a7f97ce4a80387d74b
#rb robert.srinivasiah

#ROBOMERGE-AUTHOR: jules.blok
#ROBOMERGE-SOURCE: CL 19412145 in //UE5/Release-5.0/...
#ROBOMERGE-BOT: UE5 (Release-5.0 -> Release-Engine-Staging) (v928-19376421)

[CL 19413546 by jules blok in ue5-release-engine-staging branch]