Fix for invalid D3D resource states due to storage in signed bit field
Valid D3D12_RESOURCE_STATES values now cover the the lower 31 bits.  CResourceState utilized a D3D12_RESOURCE_STATES 31 bit bit field to store a D3D12_RESOURCE_STATES.  However, the D3D12_RESOURCE_STATES enum is not explicitly typed so the compiler can interpret it as a signed type.  This allows the resource state value to overflow the supported range of the siigned 31 bit bit field rather than exactly retain the lower 31 bits as we desire.

CResourceState now stores the resource state in an unsigned 31 bit bit field.  This guarentees we retain the lower 31 bits of the resoure state exactly.  D3D12_RESOURCE_STATE_TBD and D3D12_RESOURCE_STATE_CORRUPT have been updated accordingly so that they do not utilize the top bit.  They were previously relying on the signed conversion to retain their negative values which is no longer possible.

An assertion has been added so that if a D3D12_RESOURCE_STATES is encountered that utilzes the top bit in the future, and hence cannot be correctly stored in our 31 bit bit field, we will catch it.

#rb Benjamin.Rouveyrol
#jira UE-145607
#preflight 623565d98e2a84d9bb8406a1

#ROBOMERGE-AUTHOR: eric.mcdaniel
#ROBOMERGE-SOURCE: CL 19455396 in //UE5/Main/...
#ROBOMERGE-BOT: UE5 (Main -> Release-Engine-Staging) (v933-19451510)

[CL 19455670 by eric mcdaniel in ue5-release-engine-staging branch]