Change FMath::Min/Max to fix two potholes with the current definitions
1. On custom types, they require overloads for operator <= and >= respectively; it's more idiomatic in C++ code for ordering predicates to rely solely on <
2. The current definition results in highly suboptimal code generation for floating-point types on x86 when compiling in IEEE compliant mode: https://godbolt.org/z/3zsYKaasc

In short, x86 min/max instructions are equivalent to "(a < b) ? a : b" and "(a > b) ? a : b", respectively. In the specific case where a=+0, b=-0 (IEEE signed zeros), then in comparisons +0 == -0, so we have both (a >= b) and (a <= b), but neither (a > b) nor (b < a). As a result, the original expressions for FMath::{Min,Max} return a (+0), while the changed expressions returns b (-0) in this particular case. (The a=-0, b=+0 case is analogous.)

Most code in UE currently compiles with /fp:fast or -ffast-math, which makes the compiler not respect signed zero behavior; see behavior with -fno-signed-zeros in the Compiler Explorer link. Because so much gets compiled with fast math semantics, behavior in this case was not reliable anyway. We might as well pick the definition that is efficient on x86 even without fast math.

#rb devin.doucette
#preflight 6239248304769ab4934cd0ea

#ROBOMERGE-AUTHOR: fabian.giesen
#ROBOMERGE-SOURCE: CL 19462511 in //UE5/Main/...
#ROBOMERGE-BOT: UE5 (Main -> Release-Engine-Staging) (v933-19451510)

[CL 19462521 by fabian giesen in ue5-release-engine-staging branch]