TextureCompressorModule simplify and improve parallel mip encoding
Pixel count in mip maps is in a geometric progression: each 2D mip has 4x as many pixels as the one below it,
and 3x as many as the sum of all mips below it. Therefore, a parallel for over mip levels is not very useful:
the vast majority of work is in the base mip, and leaving the base mip (which is on our critical path) for a
worker to pick up adds whatever time it takes for a worker to get started on it to our expected total time.

Therefore, change the logic so the calling thread grabs the base mip itself, and spawns exactly one other
task before to process all the other mips. This should reduce the number of tiny, useless tasks we spawn,
and because this single async task is expected to take 1/3 of the time as the base mip task, we have a decent
amount of slack for queuing delays without it showing up as a real wait time.

#rb martins.mozeiko, devin.doucette
#preflight 623b749110251d53d58911cf

#ROBOMERGE-AUTHOR: fabian.giesen
#ROBOMERGE-SOURCE: CL 19483730 in //UE5/Main/...
#ROBOMERGE-BOT: UE5 (Main -> Release-Engine-Staging) (v936-19480137)

[CL 19483769 by fabian giesen in ue5-release-engine-staging branch]