Support fallback to coarser pages in directional light SMRT - Check the page table at the ray start as a guess at what resolution is available; iterate in that clipmap - Support fallback while iterating, albeit at a performance cost
For current mapping strategies, similar performance and comparable quality to the "trilinear-like" solution today
- Resolution changes can now appear more directly at page boundaries rather than explicitly across clipmap boundaries, so depending on the given situation you may see "blocky" versions of the range bands rather than the range bands themselves. Not really a better or worse artifact, but worth noting.

NOTE: Fallbacks to very coarse data should technically also involve significant optimal biases, but we do not want to compute an optimal bias per tap. For now the texel dither optimal bias is sufficient based on the clipmap we are iterating in. We could put a minimum clamp on this but there's a direct tradeoff with losing potentially important fine-scale shadows.

#rb ola.olsson


#ROBOMERGE-AUTHOR: andrew.lauritzen
#ROBOMERGE-SOURCE: CL 19528017 via CL 19528309
#ROBOMERGE-BOT: (v937-19513599)

[CL 19528326 by andrew lauritzen in ue5-release-engine-staging branch]