Switchboard: Added "Create Config" menu item.
This automates the creation of a Switchboard config from the current map and nDisplay actor, automating this per-user drudgery. SB and SBL will be auto-launched and the config will be created and opened automatically. The user can opt to have the devices auto-connect, or also have all the ip addresses be overridden with localhost.

The user can also pick the nDisplay instance to use (when there are more than one), and select how many Editor devices he wants created. With was done w/o adding any plugin dependencies.

In addition:
* Configs can now be "Saved As" and loaded to/from anywhere on disk.
* SBL now auto-builds if necessary when launching from the Editor
* Users can now pass python scripts and arbitrary arguments for them using the --script and --scriptargs cmd line arguments, and if the script contains a subclass of SwitchboardScriptBase, its callbacks will be called at certain times of SB lifetime. In particular, during initialization. This enables certain externally driven automation of SB, and is how the automated creation of new configs is done by Unreal.

#rb zach.brockway
#jira none
#preflight 624270eac51908059a92580c

#ROBOMERGE-AUTHOR: alejandro.arango
#ROBOMERGE-SOURCE: CL 19536887 in //UE5/Main/...
#ROBOMERGE-BOT: UE5 (Main -> Release-Engine-Staging) (v937-19513599)

[CL 19536903 by alejandro arango in ue5-release-engine-staging branch]