UE5_MAIN:  Editor bug fix for mouse cursor location when tracking a cinematic camera, breaking various operations, such as the focus distance eyedropper tool or moving actors, among many other things.
Cinematic camera tracking sets a viewport that is a subset of the render target in order to achieve an aspect ratio that matches the cinematic, with letter boxing added as needed.  The mouse cursor calculations weren't taking this letterboxing into account, because they were only using the size of the viewport, and ignoring the offset of the viewport.  To minimize the risk of affecting existing code using PixelToScreen, a new utility function CursorToScreen was added, and the FViewportCursorLocation class alone was updated to use it.

#jira UE-112547
#rb brooke.hubert mihnea.balta
#rnx
#preflight 623fd7655c289ac26e3b0d1e

#ROBOMERGE-AUTHOR: jason.hoerner
#ROBOMERGE-SOURCE: CL 19583913 in //UE5/Main/...
#ROBOMERGE-BOT: UE5 (Main -> Release-Engine-Staging) (v938-19570697)

[CL 19583948 by jason hoerner in ue5-release-engine-staging branch]