UVEditor: Allow selection mode changes even if the mechanic is not currently enabled.
This adjustment is designed to support tools where the selection type before the tool is activated is not compatible with the tool's function and requires conversion (e.g. vertices to triangles). The old behavior was also partially broken, where attempts to call a mode change on an inactive selection mechanic would result in the mode being changed transparently, but with no other processing taken, such as converting the active selection or sending undo changes.

#rb	semion.piskarev
#preflight 624e9cf0a64871f1a12fe079

#ROBOMERGE-AUTHOR: nathan.mitchell
#ROBOMERGE-SOURCE: CL 19676683 in //UE5/Main/...
#ROBOMERGE-BOT: UE5 (Main -> Release-Engine-Staging) (v938-19570697)

[CL 19676729 by nathan mitchell in ue5-release-engine-staging branch]