Fix crash when changing scalability settings.
If a material instance did not have static permutation resources, the UMaterialInstance::AllMaterialsCacheResourceShadersForRendering() call in ScalabilityCVarsSinkCallback() enqueued an update uniform buffer command for that instance, instead of recreating the buffer (via CacheResourceShadersForRendering -> InitResources). However, the shader map for the parent material changes as a result of the scalability change, so the layout in the new shader map will no longer match the layout of the existing uniform buffer. The solution is to always recreate the buffer in CacheResourceShadersForRendering().

Also fixed the debug name stored in FMaterialInstanceResource. The previous code was trying to use the asset path of the owner material, but that's not set yet at the point where the instance resource is created, so the name was always that of the default material.

#jira UE-147570
#rnx
#preflight https://horde.devtools.epicgames.com/job/624aff67c61d8a458f7e4385
#rb Krzysztof.Narkowicz, Jason.Nadro

#ROBOMERGE-AUTHOR: mihnea.balta
#ROBOMERGE-SOURCE: CL 19719059 in //UE5/Main/...
#ROBOMERGE-BOT: UE5 (Main -> Release-Engine-Staging) (v938-19570697)

[CL 19719072 by mihnea balta in ue5-release-engine-staging branch]