Horde: Thread-safe store for small objects backed by a memory-mapped file, optimized for read/write performance and data integrity.
Reads are typically lock free. Writes require a lock but otherwise complete quickly and in constant time. The store is transactional, and is designed not to lose data if the process is terminated between flushes. To accomplish this, blocks are not reused until the store is flushed via a call to Save().

Blobs allocated in the store may be split into multiple cells, so do not support access to a contiguous underlying memory region. The cells comprising a blob may be enumerated and accessed individually via a handle to the blob.
The store also supports storage of tree objects which can be addressed by hash. Internally, these objects are reference counted and maintain a list of references to other blobs separately to their payload.

If the store runs out of space, supplemental stores may be allocated in granularity of <see cref="LocalBlobStoreOptions.ExpandStepSize"/>. These stores are merged into the main one during the next Save() call.

#preflight none
#preflight 625872ee7f628a9018e480e5

#ROBOMERGE-AUTHOR: ben.marsh
#ROBOMERGE-SOURCE: CL 19761303 in //UE5/Main/...
#ROBOMERGE-BOT: UE5 (Main -> Release-Engine-Staging) (v938-19570697)

[CL 19761361 by ben marsh in ue5-release-engine-staging branch]