Child Blueprint classes will now inherit assigned namespaces all the way through the native parent class hierarchy.
Also changed the UX to not show the auto-import alert unless a namespace was added into the current editor context. We technically auto-import a namespace once it's used but we don't necessarily need to tell the user about it in certain scenarios (e.g. when it's already part of the default set). The reason we do it that way is that if it is no longer present in the default set (e.g. after reparenting the child BP), the child will then display it as an explicit import.

#rnx
#jira UE-149265
#rb Benjamin.Fox
#preflight 6259e39db53385f954a62802

#ROBOMERGE-AUTHOR: phillip.kavan
#ROBOMERGE-SOURCE: CL 19778028 in //UE5/Main/...
#ROBOMERGE-BOT: UE5 (Main -> Release-Engine-Staging) (v938-19570697)

[CL 19778040 by phillip kavan in ue5-release-engine-staging branch]