Chaos Cloth: Add new HierarchicalSpatialHash acceleration structure for cloth self collision/intersections.
The HierarchicalSpatialHash can be built in parallel using a simple lock-free HashMap. Currently the HashMap uses a linked list for buckets. It might be worth considering linear probing or something similar so that bucket data is stored together in memory for queries.

In tests on various cloth assets on our current PC hardware, building the Hierarchical Spatial Hash map is over 2x faster than building the old AABBTree. Querying for the purposes of point-point proximity and edge-face intersections is within a few percent of BVH queries (but consistently slightly slower). The BVH build is serial, so speed up will depend on parallelism, but the fully serial version of the spatial hashing is not significantly slower than BVH building (this may actually speak to other issues with parallel throughput). I did try various batch sizes, but given that we typically have several thousand tris, the default batching (split evenly between all available worker threads) was the best in most cases.

The TriangleMesh proximity and intersection queries have the same interface for BVH and SpatialHash query, so calling code can be templated on which spatial acceleration structure you're using. The PBDCollisionSpringConstraints have been updated in this way to retain the previous BVH API for anyone using it.

#preflight 626c2c872660750c3f0649da
#rb kriss.gossart benn.gallagher jaco.vandyk

#ROBOMERGE-AUTHOR: alex.mcadams
#ROBOMERGE-SOURCE: CL 20045849 in //UE5/Main/...
#ROBOMERGE-BOT: UE5 (Main -> Release-Engine-Staging) (v943-19904690)

[CL 20045881 by alex mcadams in ue5-release-engine-staging branch]