This adds an option to keep trying a number of times after "Aborted sync due to errors".
This option will keep trying when the perforce server times out.  That can be useful when perforce servers are overloaded and users have bad connections and continually get kicked.  It is currently defaulted to off.  In the future we may consider defaulting this number to around 1000. The setting can be found under Options->Application Settings->Advanced->Retries on sync error.

[REVIEW] [at]Devlin.Willis [at]Spencer.Melnick [at]Daniel.Broder [at]ben.marsh
[FYI] [at]david.hamm
#jira FORT-481018
#test Pulled down entirly new branch twice.
#preflight 628be7b5e45021d097ebb1ea

#ROBOMERGE-AUTHOR: joel.anderson
#ROBOMERGE-SOURCE: CL 20337247 in //UE5/Main/...
#ROBOMERGE-BOT: UE5 (Main -> Release-Engine-Staging) (v948-20297126)

[CL 20337271 by joel anderson in ue5-release-engine-staging branch]