Shape selection in PHAT - selecting shapes in viewport selects them also in skeleton tree. This appears to have already been implemented, but had 2 bugs.
PhysicsAssetEditor::FShapeData needs all data to represent a unique shape, but did not include primitive type. Primitive index is the index into the list of primitives of a type, not into the full list of primitives, so primitive type needed to be added to this struct.

FPhysicsAssetEditor::HandleViewportSelectionChanged was looking for shapes within the list of selected constraints rather than the list of selected bodies. This change adds code to check the body selection for shapes _as well as_ the constraint selection. I left the constraint selection bit in there in case we want selecting a constraint to select the shapes it's attached to.

[FYI] benn.gallagher, chris.caulfield, gates.aldridge, satchit.subramanian


#ROBOMERGE-AUTHOR: steven.barnett
#ROBOMERGE-SOURCE: CL 20341171 via CL 20341211
#ROBOMERGE-BOT: (v948-20297126)

[CL 20341235 by steven barnett in ue5-release-engine-staging branch]