ConcertSyncClient: add a utility function for synchronizing the Asset Registry
This moves several private helper functions out of the ConcertSandboxPlatformFileUtils
namespace to instead be public in ConcertSyncClientUtil. It also adds a
SynchronizeAssetRegistry() function that takes the place of the IDirectoryWatcher::Tick()
call that was previously inserted into the sandbox discard process. This new function
can more readily be used elsewhere when file changes need to be applied immediately
in order to bring the Asset Registry up to date.

#jira UE-153112
#rb jason.walter
#preflight 628e6f44c511f67c9fd073a1

#ROBOMERGE-AUTHOR: matt.johnson
#ROBOMERGE-SOURCE: CL 20373534 in //UE5/Main/...
#ROBOMERGE-BOT: UE5 (Main -> Release-Engine-Staging) (v949-20362246)

[CL 20373545 by matt johnson in ue5-release-engine-staging branch]