nDisplay: Added ability to draw lightcards on viewport using polygons.
To use, click on the viewport "Draw LC" button, then click on the viewport at the locations where the polygon vertices should exist, then end the polygon with right mouse button. Or abort by toggling the "Draw LC" button.

The polygon is saved in an array in the light card, and is used to generate an alpha texture. This alpha texture is independent of the color texture. This alpha texture is generated dynamically, so there is no need to save it as an asset.

The default generated alpha texture dimensions is 128x128, but can be changed with DC.LightCardPolygonTextureSize.

Currently there is visual indication of the drawing process (other than the mouse becoming a cross and the Draw LC button being highlighted).
#preflight 628ec2c289000256cea3ba07

#ROBOMERGE-AUTHOR: alejandro.arango
#ROBOMERGE-SOURCE: CL 20373874 in //UE5/Main/...
#ROBOMERGE-BOT: UE5 (Main -> Release-Engine-Staging) (v949-20362246)

[CL 20373905 by alejandro arango in ue5-release-engine-staging branch]