Submitting on behalf of Jonathan.Bard, see review 19835889
Landscape: added utility function to render the merged heightmap from anywhere in world space. This is done in 2 steps : render the heightmaps  of all landscape components overlapping with the requested region into an atlas texture (uncompressed) subsection by subsection, skipping the duplicated rows/columns. Then resample the atlas at the requested location.
If the render target is 8bits/channel, the resulting heightmap will be compressed on the R and G channels. If the target is 16 bits uint, the output will contain the uncompressed height value.
The function supports heightmaps with different minimum mip level and will use the best available mip level. However, currently the intermediate atlas render target is always allocated using the landscape's full resolution.
This is meant to be compatible with runtime but is not at the moment, because it relies on ULandscapeInfo members which are not editor-only but are not available at runtime (yet). This will be fixed later on.

#rb semion.piskarev
#preflight 629007c98c23e52ef0097189


#ROBOMERGE-AUTHOR: semion.piskarev
#ROBOMERGE-SOURCE: CL 20389454 via CL 20389469
#ROBOMERGE-BOT: (v949-20362246)

[CL 20389478 by semion piskarev in ue5-release-engine-staging branch]