Sequencer: Fixed construction of child object binding models
The initial population of object binding models is precarious due to the unstructured nature of the UMovieScene data (there is no authoritative view of ancestry for objects, folders etc) and the lack of context for where objects should live.
This results in guesswork during construction for some objects (specifically, child components that are created _before_ their parent) where the complete ancestry is not known.

The bug manifested itself by child object bindings happening to be created before their parent, which would implicitly construct the parent, but leave it detached from the model hierarchy. It would then get cleaned up along with its child binding, only to be re-created in a further iteration of the construction loop, but this time without its child (which was assumed to have already been constructed).

The fix is to detect detached sub-trees on population, and always add them to the root model's children if so. This may result in object bindings being subsequently moved into folders when the folder models are created, but that is acceptable imo.

#jira UE-155939
#rb Max.Chen
#preflight 629f98de617cbe81d32f92c3

#ROBOMERGE-AUTHOR: andrew.rodham
#ROBOMERGE-SOURCE: CL 20542422 in //UE5/Main/...
#ROBOMERGE-BOT: UE5 (Main -> Release-Engine-Staging) (v954-20466795)

[CL 20542448 by andrew rodham in ue5-release-engine-staging branch]