Don't re-use actions across action sets.
If an UInputAction is part of multiple UInputMappingContext, then we need to create multiple XrAction handles. These multiple handles are still represented by a single UInputAction for blueprints and C++.

Each XrAction gets its own set of suggested bindings and can be rebound by the player separately. This allows the developer to change the bindings for UInputAction at runtime by changing which UInputMappingContext is applied.

This is an integral part of Enhanced Input and given the restrictive nature of OpenXR it will be an important feature for XR devs as well.

#rb Jeff.Fisher
#preflight 62a26bf2b17fd2cd5aa9aa67

#ROBOMERGE-AUTHOR: jules.blok
#ROBOMERGE-SOURCE: CL 20585767 in //UE5/Main/...
#ROBOMERGE-BOT: UE5 (Main -> Release-Engine-Staging) (v955-20579017)

[CL 20585781 by jules blok in ue5-release-engine-staging branch]