USDImporter: consider render context in UsdUtils::HashShadeMaterial()
This addresses an issue where UsdUtils::HashShadeMaterial() would only ever
produce a hash for the universal render context surface output of a
UsdShadeMaterial (or no hash if there was no such output). This could be
problematic if you are translating materials in any non-universal render
context (e.g. "mdl" or "mtlx") since we would produce an incorrect or empty
hash for the surface shader being translated, which could cause distinct
shaders in USD to all map incorrectly to the same material asset in Unreal.

#jira UE-142381
#rb daniel.coelho
#preflight 62a282a8cc30ea8a2849f4c4

#ROBOMERGE-AUTHOR: matt.johnson
#ROBOMERGE-SOURCE: CL 20598134 in //UE5/Main/...
#ROBOMERGE-BOT: UE5 (Main -> Release-Engine-Staging) (v955-20579017)

[CL 20598819 by matt johnson in ue5-release-engine-staging branch]