Do a pass on the alignment/padding of properties for Enhanced Input types.
UInputMappingContext -- No padding improvements, but the alignment will be better as the "Mappings" array will fill up to the 64 byte boundry

UInputAction -- Move the "Triggers" array up to the top to align at the 64 byte boundary, with the FText after it. In shipping builds, FText's are 24 bytes. This will actually result in more padding, but have it still be the same size. That means we have more room for any extra flags in the future!

FInputActionInstance -- Remove the two deprecated properties that are not in use anywhere. Seeing as we are coming out of experiemental, we should cut as much as we can now to save people from mucking stuff up in the production version. Went from 136 bytes to 96!!

FEnhancedActionKeyMapping -- Went from 144 bytes to 136 bytes. FKey is 80 bytes?! Ouch. Cut the amount of padding in half from 14 to 7

#jira UE-155097
#rb justin.hare
#rnx
#preflight 62a2dba72e1cc34f11c8aad5

#ROBOMERGE-AUTHOR: ben.hoffman
#ROBOMERGE-SOURCE: CL 20630324 in //UE5/Main/...
#ROBOMERGE-BOT: UE5 (Main -> Release-Engine-Staging) (v955-20579017)

[CL 20630380 by ben hoffman in ue5-release-engine-staging branch]