Editor configs now correctly diff against the CDO when serializing.
Previously, I had assumed that FProperty::InitializeValue() would initialize a simple member (eg. a float) to the correct value, but this appears not to be the case, rather values are zero-initialized and the CDO stores the correct default values. In the case of structs, InitializeStruct() does in fact initialize correctly.

However, this means that we can't use GetMutableDefault() to get an instance of the UObject, since then we're changing what we're diffing against and serialize no values. Changed the usages of this pattern to use singletons instead.

#jira UE-141150
[REVIEW] [at]lauren.barnes, [at]aditya.ravichandran
#preflight 62a78491f962e32e977c60e2
[FYI] lauren.barnes

#p4v-cherrypick 19919735


#ROBOMERGE-OWNER: aditya.ravichandran
#ROBOMERGE-AUTHOR: sebastian.nordgren
#ROBOMERGE-SOURCE: CL 19919719 via CL 19919729 via CL 20633551 via CL 20637295 via CL 20637521
#ROBOMERGE-BOT: (v955-20579017)

[CL 20637527 by aditya ravichandran in ue5-release-engine-staging branch]