Fix on-connection strings having different type ids.
Since there were two different events for declaring strings and FNames, one regular and one no-sync version used in on-connection callback, references would not be correctly resolved. We remove the no sync version and use only one version. In order to solve the synchronizing we defer resolving the strings in analysis.
#preflight 62a9c8932a1103ff5fb784ae

#ROBOMERGE-AUTHOR: johan.berg
#ROBOMERGE-SOURCE: CL 20668219 in //UE5/Main/...
#ROBOMERGE-BOT: UE5 (Main -> Release-Engine-Staging) (v955-20579017)

[CL 20668224 by johan berg in ue5-release-engine-staging branch]