Horde: Add AWS Lambda support for agent using the WorkerService
Adds a "one shot mode" to WorkerService that allows it to listen for exactly one lease from the server. This is a hybrid as opposed to sending the ComputeTaskMessage as part of the Lambda invocation request. Through this, the agent will listen for a lease and behave like a normal agent and share the same task queue. Functionality for starting these Lambda based agents is yet to be added.

A test task was added to aid testing in the fake HordeRpcServer.
#preflight 62b1940d0b1767c9a12480c9

#ROBOMERGE-AUTHOR: carl.bystrom
#ROBOMERGE-SOURCE: CL 20755174 in //UE5/Main/...
#ROBOMERGE-BOT: UE5 (Main -> Release-Engine-Staging) (v970-20704180)

[CL 20755178 by carl bystrom in ue5-release-engine-staging branch]