PR #9166: Ray tracing light function implementation (Contributed by ehartNV)
#jira UE-151474
#rb Yuriy.ODonnell

Create raytracing miss shaders that implement light functions for raytraced reflections, refractions and the raytracing debug primary view. Each light with a valid light function material is assigned a miss shader in the main raytracing pipeline which is used to modulate the light intensity as part of the shadowing calculation. The light function implementation intentionally converts to grayscale for visual compatibility with the deferred renderer.

As part of this CL, the default behavior for miss shaders in RT pipeline is changed. We now require the caller to use RHISetRayTracingMissShader before the first dispatch to ensure the desired shader is assigned to the desired slot in the pipeline. This relaxes some prior assumptions about the shader slot index needing to matching the shader index in the pipeline (which would not hold after having added more miss shaders to the main raytracing pipeline).

Basic validation has been added to the D3D12 RHI to ensure we are correctly binding at least the 0'th miss shader slot prior to the first dispatch.

#preflight 62bc99aca966c9f7ae8b1df9
#preflight 62bcbbf45d53ca5bcefc828f
#preflight 62bccd7a3f0d6beee247ca53

#ROBOMERGE-AUTHOR: chris.kulla
#ROBOMERGE-SOURCE: CL 20884852 in //UE5/Main/...
#ROBOMERGE-BOT: UE5 (Main -> Release-Engine-Staging) (v971-20777995)

[CL 20884876 by ehartNV in ue5-release-engine-staging branch]