Added a new RenderTrace module to perform precise ray casts using the renderer in order to determine a physical material at a specific point.
Works similar to an AsyncLineTrace except that a list of primitive components must be specified. The primitive components need a material with a UMaterialExpressionPhysicalMaterialOutput node that contains a list of UPhysicalMaterials. A draw is performed with all the primitives, and the nearest hit writes its Physical Material into the render target which is then read back from the GPU. This incurs a few frames of delay in order to not flush the render pipeline.



#ROBOMERGE-OWNER: chris.genova
#ROBOMERGE-AUTHOR: chris.genova
#ROBOMERGE-SOURCE: CL 20903194 via CL 20904515
#ROBOMERGE-BOT: (v971-20777995)

[CL 20904524 by chris genova in ue5-release-engine-staging branch]