UE-156275 - Infinite Loop due to Macros setup does not get detected, causing the Editor to freeze indefinitely on compile
Prior to tunnel&macro expansion, we need to add a step that checks for any cycles with macros. In this case, we simply perform a DFS of macro graph instances. If a cycle is detected, the BP compiler will print the cycle the path and issue an error. Failure to do so will cause the function to continuously add nodes to the source graph as macro instances are discovered, which results in an infinite loop.

#jira UE-156275
#rb phillip.kavan
#preflight 62c453bab47aed4cf1a25a43

#ROBOMERGE-AUTHOR: dave.jones2
#ROBOMERGE-SOURCE: CL 20945525 in //UE5/Main/...
#ROBOMERGE-BOT: UE5 (Main -> Release-Engine-Staging) (v971-20777995)

[CL 20945626 by dave jones2 in ue5-release-engine-staging branch]