UE-157166 - Vector literal pins lack precision with high values
Similar to other numerical Slate widgets, we need to templatize SGraphPinVector so that it can choose between a float and double representation. Currently, only the BP editor will use the double variant, while other tools continue to use a float.

This also update SGraphPinVector2D and SGraphPinVector4 to use a template. While we do use SGraphPinVector2D in BPs, we currently don't use SGraphPinVector4. However, that may change in the future.

#jira UE-157166
#preflight 62bf6898c438da7f09e95af0
#rb andrew.davidson

#ROBOMERGE-AUTHOR: dave.jones2
#ROBOMERGE-SOURCE: CL 20947777 in //UE5/Main/...
#ROBOMERGE-BOT: UE5 (Main -> Release-Engine-Staging) (v971-20777995)

[CL 20947870 by dave jones2 in ue5-release-engine-staging branch]