Fix crash after rebuilding lighting if some materials don't have the "used with static lighting" flag set.
FShaderCompilingManager::ProcessCompiledShaderMaps was storing smart pointers to material resources in a local map, which meant that they had a refcount of 2 when UMaterialInstance::CheckMaterialUsage found the missing flag and tried to recreate the resources. This in turn made the FMaterial destructor trigger an assertion failure, because the refcount was still 1 after removing the resource from the shader map. Emptying the local map before recreating the render state fixes the issue.

#jira UE-158279
#rnx
#preflight https://horde.devtools.epicgames.com/job/62c714882f2d04691854d552
#rb Jason.Nadro

#ROBOMERGE-AUTHOR: mihnea.balta
#ROBOMERGE-SOURCE: CL 21037970 in //UE5/Main/...
#ROBOMERGE-BOT: UE5 (Main -> Release-Engine-Staging) (v972-20964824)

[CL 21037979 by mihnea balta in ue5-release-engine-staging branch]