Fix manual skin cache to use render thread data only.
On the the skel mesh component is used to retrieve the mesh object. This needs to be fixed in a subsequent CL.
Skel. mesh triangles are processed section by section. This is inefficient when section are small. Ideall all section should be process at once. Also, on the the used section (i.e., having roots attached to) should be process. This should/could be addressed in future CL.

#rb none
#jira none
#preflight 62f366ffb43215cf6c16d16d

#ROBOMERGE-AUTHOR: charles.derousiers
#ROBOMERGE-SOURCE: CL 21315533 in //UE5/Main/...
#ROBOMERGE-BOT: UE5 (Main -> Release-Engine-Staging) (v972-20964824)

[CL 21315539 by charles derousiers in ue5-release-engine-staging branch]