Sequencer: Add support for multiple possessable object paste.
The previous incarnation relied on looking for the actor with the same name as the binding, which obviously isn't always the case, but was deemed good enough back then. This now will bind the pasted binding to all of the actors that the copied binding was bound to by actor name. For the situation that the actor is duplicated, it will also duplicate all of the bound actors instead of just the one.

#jira UE-160715
#preflight 62fa73d1153b17e746392d91
#rb matt.hoffman

#ROBOMERGE-AUTHOR: max.chen
#ROBOMERGE-SOURCE: CL 21406182 in //UE5/Main/...
#ROBOMERGE-BOT: UE5 (Main -> Release-Engine-Staging) (v975-21357124)

[CL 21406191 by max chen in ue5-release-engine-staging branch]