Sequencer: Fix incorrect property stats updating when upgrading floats to doubles
The "old property index" was only properly set when we encounter a property for the first time. The next time, it's already set to the new index, so we don't actually properly update the property stats.

This change keeps the old property index in the property info struct to solve this. And then it also streamlines and cleans up things a bit.

#rb andrew.rodham

[CL 21470556 by ludovic chabant in ue5-release-engine-staging branch]