nDisplay: Fix Pixel to World for orthographic projections.
PixelToWorld was always applying the root actor transform to the view matrix, but this is not correct for UV map projection.

Also fixed the RC websocket handler using hte origin component transform instead of using the rotation of its owner actor and the location of the origin component. This introduces unwanted rotations when the origin component has local rotations.

Also noticed that LC proxies don't show up in the right place when the root actor has rotations in x or y (which is unusual). This is not fixed yet.

Also noticed that Epic Stage is not placing the LCs correctly when using orthographic projection (before and after this CL). This is not fixed yet.

Note: Preflights were down.

#rb
#jira
#preflight

[CL 21476136 by alejandro arango in ue5-release-engine-staging branch]