HeterogeneousVolumes Rendering Library
This is the initial submission for the HeterogeneousVolumes rendering subsystem. The initial goal of the rendering subsystem is to formalize Niagara Fluids volume rendering as a core rendering feature. Doing so gives us the opportunity to fully integrate with the rest of the UE5 rendering feature set.

To render a Niagara Fluids volume with the HeterogeneousVolumes subsystem, the artist must modify the "Override Materials" parameter of the Mesh Renderer settings on an existing Grid3D emitter. Rather than relying on custom HLSL ray-marching nodes bound to a "Surface" Material Domain, HeterogeneousVolumes incorporates the existing Volumetric Fog paradigm. This requires the artist to bind a "Volume" Material Domain with an "Additive" Blend mode.

The HeterogeneousVolumes algorithm uses ray-marching to evaluate the direct lighting component of the Volume Rendering Equation. This is similar to the existing Niagara Fluids volume rendering algorithm, though some of its limitations have been removed. Most importantly, HeterogeneousVolumes does not impose a limit on the number of lights. HeterogeneousVolumes integrates with the Deferred Lighting model and currently supports all UE5 light types, except for the SkyLight model.

An artist can enable the HeterogeneousVolumes system with the following CVar:
r.HeterogeneousVolumes=1 (Default = 0)

The ray-marching step size can be controlled with the following CVar:
r.HeterogeneousVolumes.StepSize=X (Default = 1.0)

Shadowing fidelity can be separately controlled with the following CVar:
r.HeterogeneousVolumes.ShadowStepFactor=X (Default = 8.0)

#rb Sebastien.Hillaire, Tiago.Costa
#preflight 6301b40ffa1ac183e5c0309e

[CL 21477121 by patrick kelly in ue5-release-engine-staging branch]