[Backout] - CL21509265 [FYI] Max.Chen Original CL Desc ----------------------------------------------------------------- Animation: Add Setter function for deprecated SkeletalMesh property.
The SkeletalMesh property was deprecated and the SetSkeletalMesh function was removed. This broke Sequencer evaluation of animated SkeletalMesh properties because it could no longer find the SetSkeletalMesh function. This fixes the issue by using the new Setter functionality. A new deprecated function called SetSkeletalMesh_DEPRECATED has been added and is the Setter function for the deprecated SkeletalMesh property.

The SetSkeletalMesh_DEPRECATED function can be removed when the SkeletalMesh property is removed.

#jira UE-161336
#jira UE-161466
#preflight 6303c7a9a45b007ea247df99
#rb ludovic.chabant, josie.yang, thomas.sarkanen

[CL 21513758 by max chen in ue5-release-engine-staging branch]