Sequencer: handle the case where multiple ECS entities are driving a single transform section
When auto-keying, we try to figure out the contribution of the section being keyed compared to the overall current transform value (for when that transform is computed from multiple sections, like an absolute plus an additive). This code was working under the assumption that a transform section's contribution is built from a single ECS entity. This isn't a valid assumption anymore with overridable channels.

If a transform section has some overriden channels, there will be one extra ECS entity for each overriden channel. So the recomposition code now needs to deal with multiple entities, which this change addresses.

Note that the index of channels can now vary: if Translation.X is overriden, the first "double" channel will be Translation.Y, with index 0 in the bucket of channels of type "double". But Translation.Y is really channel 1, not channel 0. So this change also makes use of GetChannelBySortOrder, instead of GetChannel, so that we get the correct channel to key.

#rb max.chen
#jira none
#preflight 63053492c85b7fef228cf26b

[CL 21531707 by ludovic chabant in ue5-release-engine-staging branch]