Add support for resizing the scene rendertargets by dynamically adjusting the max resolution based on a quantized version of the computed resolution fraction. This results in more efficient clears and resolves (since these now scale with resolution), and more efficient use of fast VRAM on platforms which use it.
Quantization can be controlled via r.DynamicRes.UpperBoundQuantization, which specifies the number of steps. The step size is given by (MaxScreenPercentage-MinScreenPercentage)/UpperBoundQuantization. Set this to 0 to disable.

Only recommended for use with the transient allocator (on supported platforms) with a large transient texture cache (e.g RHI.TransientAllocator.TextureCacheSize=512). A rule-of-thumb for cache size is 32 * r.DynamicRes.UpperBoundQuantization.

#rb guillaume.abadie

[CL 21571995 by ben woodhouse in ue5-release-engine-staging branch]